Remove flux store usages from ChannelIdentifierRouter
